+++ cmd/describe/describe-consumer-group.go
@@ -6,9 +6,9 @@ import (
 	"github.com/spf13/cobra"
 )
 
-var consumerGroupFlags consumergroups.DescribeConsumerGroupFlags
-
 func newDescribeConsumerGroupCmd() *cobra.Command {
+
+	var flags consumergroups.DescribeConsumerGroupFlags
 
 	var cmdDescribeConsumerGroup = &cobra.Command{
 		Use:     "consumer-group GROUP",
@@ -16,17 +16,17 @@ func newDescribeConsumerGroupCmd() *cobra.Command {
 		Short:   "describe a consumerGroup",
 		Args:    cobra.MinimumNArgs(1),
 		Run: func(cmd *cobra.Command, args []string) {
-			if err := (&consumergroups.ConsumerGroupOperation{}).DescribeConsumerGroup(consumerGroupFlags, args[0]); err != nil {
+			if err := (&consumergroups.ConsumerGroupOperation{}).DescribeConsumerGroup(flags, args[0]); err != nil {
 				output.Fail(err)
 			}
 		},
 	}
 
-	cmdDescribeConsumerGroup.Flags().BoolVarP(&consumerGroupFlags.OnlyPartitionsWithLag, "only-with-lag", "l", false, "show only partitions that have a lag")
-	cmdDescribeConsumerGroup.Flags().StringVarP(&consumerGroupFlags.FilterTopic, "topic", "t", "", "show group details for given topic only")
-	cmdDescribeConsumerGroup.Flags().StringVarP(&consumerGroupFlags.OutputFormat, "output", "o", flags.OutputFormat, "output format. One of: json|yaml|wide")
-	cmdDescribeConsumerGroup.Flags().BoolVarP(&consumerGroupFlags.PrintTopics, "print-topics", "T", true, "print topic details")
-	cmdDescribeConsumerGroup.Flags().BoolVarP(&consumerGroupFlags.PrintMembers, "print-members", "m", true, "print group members")
+	cmdDescribeConsumerGroup.Flags().BoolVarP(&flags.OnlyPartitionsWithLag, "only-with-lag", "l", false, "show only partitions that have a lag")
+	cmdDescribeConsumerGroup.Flags().StringVarP(&flags.FilterTopic, "topic", "t", "", "show group details for given topic only")
+	cmdDescribeConsumerGroup.Flags().StringVarP(&flags.OutputFormat, "output", "o", flags.OutputFormat, "output format. One of: json|yaml|wide")
+	cmdDescribeConsumerGroup.Flags().BoolVarP(&flags.PrintTopics, "print-topics", "T", true, "print topic details")
+	cmdDescribeConsumerGroup.Flags().BoolVarP(&flags.PrintMembers, "print-members", "m", true, "print group members")
 
 	return cmdDescribeConsumerGroup
 }