+++ cmd/get/get-consumer-groups.go
@@ -6,9 +6,9 @@ import (
 	"github.com/spf13/cobra"
 )
 
-var consumerGroupFlags consumergroups.GetConsumerGroupFlags
-
 func newGetConsumerGroupsCmd() *cobra.Command {
+
+	var flags consumergroups.GetConsumerGroupFlags
 
 	var cmdGetConsumerGroups = &cobra.Command{
 		Use:     "consumer-groups",
@@ -16,14 +16,14 @@ func newGetConsumerGroupsCmd() *cobra.Command {
 		Short:   "list available consumerGroups",
 		Args:    cobra.MaximumNArgs(0),
 		Run: func(cmd *cobra.Command, args []string) {
-			if err := (&consumergroups.ConsumerGroupOperation{}).GetConsumerGroups(consumerGroupFlags); err != nil {
+			if err := (&consumergroups.ConsumerGroupOperation{}).GetConsumerGroups(flags); err != nil {
 				output.Fail(err)
 			}
 		},
 	}
 
-	cmdGetConsumerGroups.Flags().StringVarP(&consumerGroupFlags.OutputFormat, "output", "o", consumerGroupFlags.OutputFormat, "output format. One of: json|yaml|wide|compact")
-	cmdGetConsumerGroups.Flags().StringVarP(&consumerGroupFlags.FilterTopic, "topic", "t", "", "show groups for given topic only")
+	cmdGetConsumerGroups.Flags().StringVarP(&flags.OutputFormat, "output", "o", flags.OutputFormat, "output format. One of: json|yaml|wide|compact")
+	cmdGetConsumerGroups.Flags().StringVarP(&flags.FilterTopic, "topic", "t", "", "show groups for given topic only")
 
 	return cmdGetConsumerGroups
 }