+++ operations/producer/MessageSerializer.go
@@ -1,6 +1,8 @@
 package producer
 
 import (
+	"encoding/base64"
+	"encoding/hex"
 	"github.com/Shopify/sarama"
 	"github.com/pkg/errors"
 	"regexp"
@@ -10,6 +12,11 @@ type MessageSerializer interface {
 	CanSerialize(topic string) (bool, error)
 	Serialize(key, value []byte, flags ProducerFlags) (*sarama.ProducerMessage, error)
 }
+
+const (
+	HEX    = "hex"
+	BASE64 = "base64"
+)
 
 func parseHeader(raw string) (key, value string, err error) {
 
@@ -44,4 +51,38 @@ func createRecordHeaders(flags ProducerFlags) ([]sarama.RecordHeader, error) {
 		recordHeaders[i].Value = sarama.ByteEncoder(value)
 	}
 	return recordHeaders, nil
+}
+
+func decodeBytes(data []byte, encoding string) ([]byte, error) {
+	if data == nil {
+		return nil, nil
+	}
+
+	var out []byte
+	switch encoding {
+	case HEX:
+		out = make([]byte, hex.DecodedLen(len(data)))
+		if _, err := hex.Decode(out, data); err != nil {
+			return nil, err
+		}
+		return out, nil
+	case BASE64:
+		out = make([]byte, base64.StdEncoding.DecodedLen(len(data)))
+		if _, err := base64.StdEncoding.Decode(out, data); err != nil {
+			return nil, err
+		}
+		return out[:clen(out)], nil
+	default:
+		return data, nil
+	}
+}
+
+// https://stackoverflow.com/a/27834860/12143351
+func clen(n []byte) int {
+	for i := 0; i < len(n); i++ {
+		if n[i] == 0 {
+			return i
+		}
+	}
+	return len(n)
 }